+++ packages/fedify/src/runtime/langstr.ts
@@ -1,21 +1,23 @@
-import { type LanguageTag, parseLanguageTag } from "@phensley/language-tag";
-
 /**
  * A language-tagged string which corresponds to the `rdf:langString` type.
  */
 export class LanguageString extends String {
-  readonly language: LanguageTag;
+  /**
+   * The locale of the string.
+   * @since 2.0.0
+   */
+  readonly locale: Intl.Locale;
 
   /**
    * Constructs a new `LanguageString`.
    * @param value A string value written in the given language.
-   * @param language The language of the string.  If a string is given, it will
-   *                 be parsed as a `LanguageTag`.
+   * @param locale The language of the string.  If a string is given, it will
+   *               be parsed as a `Intl.Locale` object.
    */
-  constructor(value: string, language: LanguageTag | string) {
+  constructor(value: string, language: Intl.Locale | string) {
     super(value);
-    this.language = typeof language === "string"
-      ? parseLanguageTag(language)
+    this.locale = typeof language === "string"
+      ? new Intl.Locale(language)
       : language;
   }
 
@@ -23,7 +25,7 @@ export class LanguageString extends String {
     inspect: typeof Deno.inspect,
     options: Deno.InspectOptions,
   ): string {
-    return `<${this.language.compact()}> ${inspect(this.toString(), options)}`;
+    return `<${this.locale.baseName}> ${inspect(this.toString(), options)}`;
   }
 
   [Symbol.for("nodejs.util.inspect.custom")](
@@ -31,6 +33,6 @@ export class LanguageString extends String {
     options: unknown,
     inspect: (value: unknown, options: unknown) => string,
   ): string {
-    return `<${this.language.compact()}> ${inspect(this.toString(), options)}`;
+    return `<${this.locale.baseName}> ${inspect(this.toString(), options)}`;
   }
 }