+++ packages/fedify/src/vocab/vocab.test.ts
@@ -1,4 +1,3 @@
-import { parseLanguageTag } from "@phensley/language-tag";
 import {
   assert,
   assertEquals,
@@ -890,7 +889,7 @@ const sampleValues: Record<string, any> = {
     0x30, 0x65, 0x59, 0x76, 0xf0, 0xb0, 0xdb, 0x78, 0x0d, 0xf5, 0xc1, 0xad,
     0x3f, 0xbd, 0xf3, 0x07,
   ]),
-  "fedify:langTag": parseLanguageTag("en"),
+  "fedify:langTag": new Intl.Locale("en-Latn-US"),
   "fedify:url": new URL("https://fedify.dev/"),
   "fedify:publicKey": rsaPublicKey1.publicKey,
   "fedify:multibaseKey": ed25519PublicKey.publicKey,